Make examples work with current and released dependencies.
The goal of this commit is:

- Make the examples use the most recent version of our starter during development on the starter and the example
- Make the examples usable standalone for the user
- When used standalone, the example shall use a released version of the starter

This is achivied by adding default properties for revision, sha1 and changelist to all the examples, wrapped in a profile each.

The profiles will get activated when there is no such property. That is the case, when either the surrounding project (not the parent, as the examples are independent of the main project) is build without CI setting those or when a user wants to checked out the example alone.

The three properties then will be concatenated into one neo4j-java-driver-spring-boot-starter.version property in the same format as the surrounding project defines.

Note that all of the examples now use 999-SNAPSHOT as their version to make it perfectly clear that they are not part of the release or versioning process of the main artifact.