GH-1 - Configure logging to go through spring-jcl.
This removes the configuration properties for the drivers logging class and the corresponding log level and adds a small adapter to Springs JCL Logging.

As the driver uses String.format for logging, we should revisit this on Spring Boot 2.2 with the new JCL Log Accessor and LogMessage support coming in Spring 5.2.

This fixes #1.