GH-15 - Add dedicated support for the Neo4j Test Harness.
This brings in a new module, `neo4j-java-driver-test-harness-spring-boot-autoconfigure`, that recognises both Neo4j 3.5 test harness (the default) and Neo4j 4.0 test harness. 

If it detects the test harness on the classpath and - if there's no explicitly configured driver instance - creates an instance of a testharness if there's none and an instance of the driver configured to use the new test harness or an existing test harness instance.

The module defaults to 3.5.x so that it can stay on JDK 8.  The examples - which had been JDK 11+ already - change the dependency to 4.0, so that there are tests in place for both scenarios. 

The module comes with documentation in form of the examples.